test : 주문 서비스 테스트 설명

+++ src/main/java/kitchenpos/domain/Product.java
@@ -63,8 +63,8 @@ public class Product {
             return this;
         }
 
-        public Builder price(BigDecimal price) {
-            this.price = price;
+        public Builder price(long price) {
+            this.price = BigDecimal.valueOf(price);
             return this;
         }
